Refactor and add diff
