+++ src/basic_statistics.rs
@@ -337,78 +337,6 @@ pub mod stats {
         }
     }
 
-    /// Returns the approximate integral of the curve defined by `y` and `x` using the trapezoidal method.
-    /// ```typescript
-    /// let y = [1.0, 1.5, 2.0];
-    /// let x = [1.0, 2.0, 3.0];
-    /// let A = trapz(x, y);
-    /// assert_eq(A, 3.0);
-    /// ```
-    /// ```typescript
-    /// let y = [1, 2, 3];
-    /// let x = [1, 2, 3];
-    /// let A = trapz(x, y);
-    /// assert_eq(A, 4.0);
-    /// ```
-    #[rhai_fn(name = "trapz", return_raw)]
-    pub fn trapz(x: Array, y: Array) -> Result<Dynamic, Box<EvalAltResult>> {
-        if x.len() != y.len() {
-            return Err(EvalAltResult::ErrorArithmetic(
-                format!("The arrays must have the same length"),
-                Position::NONE,
-            )
-            .into());
-        }
-
-        // Convert if needed
-        let mut X: Vec<FLOAT> = if x[0].is::<INT>() {
-            x.iter().map(|el| el.as_int().unwrap() as FLOAT).collect()
-        } else {
-            x.iter().map(|el| el.as_float().unwrap()).collect()
-        };
-
-        // Convert if needed
-        let mut Y: Vec<FLOAT> = if y[0].is::<INT>() {
-            y.iter().map(|el| el.as_int().unwrap() as FLOAT).collect()
-        } else {
-            y.iter().map(|el| el.as_float().unwrap()).collect()
-        };
-
-        let mut trapsum = 0.0;
-        for i in 1..x.len() {
-            trapsum += (Y[i] + Y[i - 1]) * (X[i] - X[i - 1]) / 2.0;
-        }
-        Ok(Dynamic::from_float(trapsum))
-    }
-
-    /// Returns the approximate integral of the curve defined by `y` using the trapezoidal method.
-    /// Assumes that x-values have unit spacing.
-    /// ```typescript
-    /// let y = [1.0, 1.5, 2.0];
-    /// let A = trapz(y);
-    /// assert_eq(A, 3.0);
-    /// ```
-    /// ```typescript
-    /// let y = [1, 2, 3];
-    /// let A = trapz(y);
-    /// assert_eq(A, 4.0);
-    /// ```
-    #[rhai_fn(name = "trapz", return_raw)]
-    pub fn trapz_unit(y: Array) -> Result<Dynamic, Box<EvalAltResult>> {
-        // Convert if needed
-        let mut Y: Vec<FLOAT> = if y[0].is::<INT>() {
-            y.iter().map(|el| el.as_int().unwrap() as FLOAT).collect()
-        } else {
-            y.iter().map(|el| el.as_float().unwrap()).collect()
-        };
-
-        let mut trapsum = 0.0;
-        for i in 1..y.len() {
-            trapsum += (Y[i] + Y[i - 1]) / 2.0;
-        }
-        Ok(Dynamic::from_float(trapsum))
-    }
-
     /// Returns the variance of a 1-D array.
     /// ```typescript
     /// let data = [1, 2, 3];
